watch he position of tag, under the first loop, not at a global space (#8)
* watch he position of tag, under the first loop, not at a global space
learn how to debug, how to check a recursion by printing num to check where the recursion stoped, why it was stopped
It was stopped because the programme thought it was a leaf node.
why it was a leaf node, because tag == 1, the loop did not do anything. The recursion line(queen8(num - 1, new_board))
was not excuted, so it was thought as a leaf node and go back to the upper level.
learn to use import copy at the begging, because we need to use new_board as a copy to show different board.
For this problem, it has 8 new_boards at most.

* 8queen
It is a recursive problem, need to be more abstract. do more functions!

* 8queen
make the three directions in a function.

* 8queen
changed the function name